Refactor DataContext to use customised react hook
